[FIX] Memperbaiki tipe data tabel pesanan

+++ database/migrations/2024_06_05_150634_create_pesanan_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained('users');
             $table->foreignId('produk_id')->constrained('produk');
-            $table->float('jumlah');
-            $table->float('total');
+            $table->integer('jumlah');
+            $table->integer('total');
             $table->string('alamat');
             $table->enum('status', ['Menunggu Konfirmasi', 'Sedang Dikirim', 'Selesai', 'Dibatalkan'])->default('Menunggu Konfirmasi');
             $table->mediumText('catatan')->nullable();
